+++ app/test/page.tsx
@@ -1,16 +1,18 @@
+/**
+ * Dev/test error-style page with contact links.
+ * Notes: docs/notes/app/test-page.md
+ */
 "use client";
 
-// need to mark it as useClient, because of the way Next passes errors to this component Error: Functions cannot be passed directly to Client Components unless you explicitly expose it by marking it with "use server". <... buildId=... assetPrefix="" initialCanonicalUrl=... initialTree=... initialHead=... globalErrorComponent={function} children=...>
 import Link from "next/link";
 import PageTitleWithImages from "@components/ReusableSmallComponents/TitlesOrHeadings/PageTitleWithImages";
 import ListWithPawPrintIcon from "@components/ReusableSmallComponents/ListWithPawPrintIcon";
-
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faEnvelope, faMessage } from "@fortawesome/free-solid-svg-icons";
 import PawPrintIcon from "@components/ReusableSmallComponents/iconsOrSvgImages/PawPrintIcon";
 import Image from "next/image";
 
-export default function CustomError() {
+export default function TestPage() {
   return (
     <>
       <PageTitleWithImages